requirements.txt
bedrock-agentcore>=1.0.0
bedrock-agentcore-starter-toolkit
strands-agents>=0.1.0
boto3>=1.34.0
